backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
beautifulsoup4==4.13.4
lxml==5.4.0
python-dotenv==1.1.0
anthropic==0.52.0
pydantic==2.11.3
